fetching ion wallet txs and balance from indexer, ice name fix (#345)

+++ cmd/heimdall-identity-io/wallet_views.go
@@ -233,7 +233,7 @@ func (s *service) ModifyWalletView(
 	return server.OK[WalletView](view), nil
 }
 
-func withPagination(ctx context.Context, token string, limit uint) context.Context {
+func withPagination(ctx context.Context, token string, limit uint64) context.Context {
 	ctx = context.WithValue(ctx, "paginationToken", token)
 	ctx = context.WithValue(ctx, "paginationLimit", limit)
 	return ctx
